api/requirements.txt
fastapi==0.99.1
uvicorn==0.22.0
httpx==0.24.1
python-dotenv==1.0.0
pydantic==1.10.8